Fail closed on invalid Puzzletron execution records
